Only write if file does not exist

+++ shares.py
@@ -146,11 +146,12 @@ def addShare(baseDir: str, \
 
     # indicate that a new share is available
     newShareFile=baseDir+'/accounts/.newShare'
-    try:
-        with open(newShareFile, 'w') as fp:
-            fp.write('\n')
-    except:
-        pass
+    if not os.path.isfile(newShareFile):
+        try:
+            with open(newShareFile, 'w') as fp:
+                fp.write('\n')
+        except:
+            pass
 
 def expireShares(baseDir: str) -> None:
     """Removes expired items from shares
